+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed release engine not respecting `-g false` (do not release Globals)
 - Fixed column order in DQE results graph sometimes resulting in shifted colors (e.g. Correct appearing in red instead of green)
 - Fixed Prediction rules never being run when value being considered is null (DQE).
+- Fixed a bug creating a cohort without specifying a Project from the console
 
 ### Added